feat: golang test general
multi problems tests generator
