+++ python/constants/__init__.py
@@ -8,6 +8,6 @@ from .submit_query import (RECENT_SUBMISSIONS_QUERY, RECENT_AC_SUBMISSIONS_QUERY
 from .user_query import USER_PROFILE_PUBLIC_QUERY
 from .code_templates import (TESTCASE_TEMPLATE_PYTHON, TESTCASE_TEMPLATE_PYTHON_TESTCASES,
                              SOLUTION_TEMPLATE_PYTHON, SOLUTION_TEMPLATE_GOLANG,
-                             SOLUTION_TEMPLATE_GOLANG_MODIFY_IN_PLACE,
+                             SOLUTION_TEMPLATE_GOLANG_MODIFY_IN_PLACE, TESTCASE_TEMPLATE_GOLANG,
                              SOLUTION_TEMPLATE_JAVA, SOLUTION_TEMPLATE_CPP, SOLUTION_TEMPLATE_TYPESCRIPT)
 from .display import (SUBMIT_BASIC_RESULT, SUBMIT_SUCCESS_RESULT, SUBMIT_FAIL_RESULT)
